Update 7/2/24
Create a van: tested and fully working.
Update a van: Need form implementation. Reusability of same form increased complexity and error prone. Design a dedicated moda to then assess the logic and review.
Delete a van: bugged. It deletes the document but it returns a fail to the UI and a refresh of the browser is necessary to view the update of the page.